change ConditionCountMetricSpec to specialized_resolvers.py
